// packages/runtime/src/old-slate-types.ts
export interface ValueJSON {
  object?: 'value' | undefined
  annotations?: { [key: string]: AnnotationJSON } | undefined
  data?: { [key: string]: any } | undefined
  document?: DocumentJSON | undefined
  selection?: SelectionJSON | undefined
}
// ...
export interface DocumentJSON {
  object?: 'document' | undefined
  nodes?: NodeJSON[] | undefined
  key?: string | undefined
  data?: { [key: string]: any } | undefined
}
// ...
export type NodeJSON = DocumentJSON | BlockJSON | InlineJSON | TextJSON
// ...
export interface MarkJSON {
  object?: 'mark' | undefined
  type: string
  data?: { [key: string]: any } | undefined
}
// ...
export interface SelectionJSON {
  object?: 'selection' | undefined
  anchor?: PointJSON | undefined
  focus?: PointJSON | undefined
  isFocused?: boolean | undefined
  marks?: MarkJSON[] | undefined
}
// ...
export interface AnnotationJSON {
  object?: 'annotation' | undefined
  key: string
  type: string
  data?: { [key: string]: any } | undefined
  anchor?: PointJSON | undefined
  focus?: PointJSON | undefined
}
// ...
export interface PointJSON {
  object?: 'point' | undefined
  key?: string | undefined
  offset?: number | undefined
  path?: number[] | undefined
}